// frontend/src/components/chat/DirectMessageCard.tsx
import type { Conversation } from "@/types/chat";
import ChatCard from "./ChatCard";
import { useAuthStore } from "@/stores/useAuthStore";
import { useChatStore } from "@/stores/useStateStore";
import { cn } from "@/lib/utils";
import UserAvatar from "./UserAvatar";
import StatusBadge from "./StatusBadge";
import UnReadCountBadge from "./UnReadCountBadge";

const DirectMessageCard = ({ convo }: { convo: Conversation }) => {
  const { user } = useAuthStore();
  const { activeConversationId, setActiveConversation, messages ,fetchMessages} =
    useChatStore();
  if (!user) return null;

  const ortherUser = convo.participants.find((p) => p._id !== user._id);
  if (!ortherUser) return null;

  const unreadCount = convo.unreadCounts[user._id];

  const lastMessage = convo.lastMessage?.content ?? " ";

  const handleSelectConversation = async (id: string) => {
    setActiveConversation(id);
    if (!messages[id]) {
      await fetchMessages()
    }
  };

  return (
    <ChatCard
      convoId={convo._id}
      name={ortherUser.displayName ?? ""}
      timestamp={
        convo.lastMessage?.createdAt
          ? new Date(convo.lastMessage.createdAt)
          : undefined
      }
      isActive={activeConversationId === convo._id}
      onSelect={handleSelectConversation}
      unreadCount={unreadCount}
      leftSection={
        <>
          <UserAvatar
            type="sidebar"
            name={ortherUser.displayName ?? ""}
            avatarUrl={ortherUser.avatarUrl ?? undefined}
          />
          <StatusBadge status="offline" />
          {
            unreadCount>0&&<UnReadCountBadge unreadCount={unreadCount} />
          }
        </>
      }
      subtitle={
        <p
          className={cn(
            "text-sm truncate",
            unreadCount > 0
              ? "font-medium text-foreground"
              : "text-muted-foreground",
          )}
        >
          {lastMessage}
        </p>
      }
    />
  );
};

export default DirectMessageCard;
